Refactored popover into own component
